requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.0
future==0.18.2
h11==0.9.0
httpcore==0.10.2
httptools==0.1.1
httpx==0.14.2
idna==2.10
pydantic==1.6.1
pyngrok==4.1.10
python-dotenv==0.14.0
python-multipart==0.0.5
PyYAML==5.3.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
starlette==0.13.6
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
